sek-17: rename
