+++ packages/client/src/pages/Forum/ForumTopic/components/Comment/Comment.module.scss
@@ -1,0 +1,18 @@
+.text.text {
+  width: 100%;
+}
+
+.reply {
+  padding: 0 16px;
+}
+
+.info {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.author {
+  padding: 0 16px;
+}
